Move tag groups from "defaults" to "external"
